All spawn priority to be a double. Also bug fixes and tweaks.

+++ src/main/java/com/github/zarena/entities/ZEntityType.java
@@ -10,5 +10,5 @@ public interface ZEntityType extends EntityType
 	
 	public int getMinimumSpawnWave();
 	
-	public int getSpawnPriority();
+	public double getSpawnPriority();
 }
